requirements.txt
Flask==2.3.2
google_api_python_client==2.86.0
google_auth_oauthlib==1.0.0
protobuf==4.23.1
python-dotenv==1.0.0
telepot==12.7
watchdog==3.0.0
gunicorn